requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
dotenv==0.9.9
exceptiongroup==1.3.0
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
psycopg2-binary==2.9.10
Pygments==2.19.1
pytest==8.4.0
python-dotenv==1.1.0
requests==2.32.3
tomli==2.2.1
typing_extensions==4.14.0
urllib3==2.4.0
